GSK-1191 Moved tests transformation to different package
